+++ README.md
@@ -5,13 +5,13 @@
 [![bithound](https://www.bithound.io/github/Kronos-Integration/kronos-service-registry/badges/score.svg)](https://www.bithound.io/github/Kronos-Integration/kronos-service-registry)
 [![codecov.io](http://codecov.io/github/Kronos-Integration/kronos-service-registry/coverage.svg?branch=master)](http://codecov.io/github/Kronos-Integration/kronos-service-registry?branch=master)
 [![Coverage Status](https://coveralls.io/repos/Kronos-Integration/kronos-service-registry/badge.svg)](https://coveralls.io/r/Kronos-Integration/kronos-service-registry)
-[![Code Climate](https://codeclimate.com/github/Kronos-Integration/kronos-service-registry/badges/gpa.svg)](https://codeclimate.com/github/Kronos-Integration/kronos-service-registry)
 [![Known Vulnerabilities](https://snyk.io/test/github/Kronos-Integration/kronos-service-registry/badge.svg)](https://snyk.io/test/github/Kronos-Integration/kronos-service-registry)
 [![GitHub Issues](https://img.shields.io/github/issues/Kronos-Integration/kronos-service-registry.svg?style=flat-square)](https://github.com/Kronos-Integration/kronos-service-registry/issues)
 [![Stories in Ready](https://badge.waffle.io/Kronos-Integration/kronos-service-registry.svg?label=ready&title=Ready)](http://waffle.io/Kronos-Integration/kronos-service-registry)
 [![Dependency Status](https://david-dm.org/Kronos-Integration/kronos-service-registry.svg)](https://david-dm.org/Kronos-Integration/kronos-service-registry)
 [![devDependency Status](https://david-dm.org/Kronos-Integration/kronos-service-registry/dev-status.svg)](https://david-dm.org/Kronos-Integration/kronos-service-registry#info=devDependencies)
 [![docs](http://inch-ci.org/github/Kronos-Integration/kronos-service-registry.svg?branch=master)](http://inch-ci.org/github/Kronos-Integration/kronos-service-registry)
+[![XO code style](https://img.shields.io/badge/code_style-XO-5ed9c7.svg)](https://github.com/sindresorhus/xo)
 [![downloads](http://img.shields.io/npm/dm/kronos-service-registry.svg?style=flat-square)](https://npmjs.org/package/kronos-service-registry)
 [![Commitizen friendly](https://img.shields.io/badge/commitizen-friendly-brightgreen.svg)](http://commitizen.github.io/cz-cli/)
