+++ doc/jsdoc.json
@@ -1,0 +1,6 @@
+{
+  "plugins": ["node_modules/jsdoc-babel"],
+  "babel": {
+    "presets": ["env"]
+  }
+}
